firmware_requirements.txt
numpy==1.19.0
pandas==1.1.3
picamera==1.13
pkg-resources==0.0.0
pyserial==3.4
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
